jros2 conversion

+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/behaviorTree/BehaviorTreeRootNodeState.java
@@ -1,6 +1,6 @@
 package us.ihmc.behaviors.behaviorTree;
 
-import behavior_msgs.msg.dds.BehaviorTreeRootNodeStateMessage;
+import behavior_msgs.BehaviorTreeRootNodeStateMessage;
 import gnu.trove.map.hash.TLongObjectHashMap;
 import org.apache.commons.lang3.mutable.MutableInt;
 import us.ihmc.avatar.drcRobot.DRCRobotModel;
@@ -99,7 +99,7 @@ public class BehaviorTreeRootNodeState extends BehaviorTreeNodeState<BehaviorTre
       super.toMessage(message.getState());
       scene.toMessage(message.getScene());
       message.setAutomaticExecution(automaticExecution.toMessage());
-      message.setExecutionNextIndex(executionNextIndex.toMessage());
+      message.setExecutionNextIndex((short) executionNextIndex.toMessage());
       message.setManualExecutionRequested(manualExecutionRequested.toMessage());
       message.setConcurrencyEnabled(concurrencyEnabled.toMessage());
       message.setPreviewModeEnabled(previewModeEnabled.toMessage());
